feat: Adds synchronization byte and CatSniffer instructions
Adds instructions for flashing the CatSniffer with the LoRa Sniffer firmware.
Also adds the SYNC_WORLD byte and includes it in the transmit sequence to the CatSniffer.
Only decrypts and shows details when the received data contains "@S".
The transmit commands to the sniffer are also terminated by a newline character.